Integration test for prediction task with fail scenario
